requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
pyyaml==6.0.1
streamlit==1.36.0
